chachara/requirements.txt
flask
Flask-Negotiate
feedformatter
feedparser
PubSubHubbub_Publisher
